Professors can now delete archived projects

+++ src/main/java/sysc4806/project/demo/ProjectController.java
@@ -14,6 +14,8 @@ public class ProjectController {
     @Autowired
     private StudentRepository studentRepo;
 
+    @Autowired ProfessorRepository profRepo;
+
     @PostMapping("/project/createProject")
     public Project newProject(@RequestBody Project proj){
         return projectRepo.save(proj);
@@ -26,7 +28,24 @@ public class ProjectController {
 
     @DeleteMapping("/project/delete")
     public void deleteProject(@RequestParam Long id){
-        projectRepo.deleteById(id);
+        if (projectRepo.findById(id).isPresent()) {
+            Project target = projectRepo.findById(id).get();
+            List<Student> studentsInProj = studentRepo.findByProject(target);
+
+            for (Student stud : studentsInProj){
+                stud.setProject(null);
+                studentRepo.save(stud);
+            }
+
+            if (target.getProfessor() != null){
+                Professor prof = target.getProfessor();
+                prof.removeProject(target);
+                profRepo.save(prof);
+            }
+
+            projectRepo.deleteById(id);
+        }
+
     }
 
     @PutMapping("/project/addRestriction")
